# docs/swagger-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API",
        "description": "API",
        "version": "1.0.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/auth/verify-otp": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Verify OTP",
                "description": "Verify OTP",
                "operationId": "verifyOTP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Verify OTP",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "example": "example01@gmail.com"
                                },
                                "otp": {
                                    "example": "knDAfD"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "example": "logged in"
                                },
                                "status": {
                                    "example": "true"
                                },
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "username": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "no_telp": {
                                            "type": "string"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}